skills_use/pyproject.toml
[project]
name = "skills-use"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.3.7",
    "dotenv>=0.9.9",
    "langchain-openai>=1.1.7",
    "rich>=14.2.0",
]
